Make shortcuts better to use
For application switching we use the left ALT key plus letters from the left side of the keyboard so the right hand can stay on the mouse.

For Wezterm the LEADER was moved to `,` so it is easier to press.
In addition the concept of key tables was introduced plus better key selection for certain actions.